implementing firebase db into app.js and create account page pt1
