+++ src/components/App/App.js
@@ -7,14 +7,27 @@ import StartPage from './../StartPage';
 import PostingPage from './../PostingPage';
 import CreateAccountPage from './../CreateAccountPage';
 import SignInPage from './../SignInPage';
-import * as firebase from 'firebase';
+import firebase from 'firebase';
 
 /**
  * App Component
  * @type {Class}
  */
-export default class App extends React.Component {
+class App extends React.Component {
 
+  // Initialize Firebase
+  constructor(props) {
+    super(props);
+    var config = {
+      apiKey: "AIzaSyC5QbpGLGC2MxpXn8ODBDWcXCPl3mFdG-o",
+      authDomain: "postpin-c9e80.firebaseapp.com",
+      databaseURL: "https://postpin-c9e80.firebaseio.com",
+      projectId: "postpin-c9e80",
+      storageBucket: "postpin-c9e80.appspot.com",
+      messagingSenderId: "172723892563"
+    };
+    firebase.initializeApp(config);
+  }
 
   /**
    * Render function for App Component
@@ -35,7 +48,8 @@ export default class App extends React.Component {
               <PostingPage />
             )} />
             <Route exact={true} path="/create-account" render={()=> (
-              <CreateAccountPage />
+              //<CreateAccountPage />
+              <CreateAccountPage db ={firebase} />
             )} />
             <Route exact={true} path="/sign-in" render={()=> (
               <SignInPage />
@@ -46,3 +60,5 @@ export default class App extends React.Component {
     );
   }
 }
+
+export default App;
